feat: Refine remote control and process monitoring
This commit introduces targeted enhancements to the remote control and process monitoring features of OmniSync.

- **Remote Control UX & Performance:**
  - Implemented interpolated mouse movement on the Hub for a significantly smoother cursor experience on the client.
  - Adjusted mouse sensitivity on the Android remote control for more precise control.
  - Refined the remote control UI on Android, improving layout and button sizing for better ergonomics.
  - Updated icons to use `AutoMirrored` for better right-to-left (RTL) language support.

- **Process Monitoring Data:**
  - Enriched the `ProcessInfo` data model to include `MemoryUsage` and `CpuUsage`, providing more detailed process information to clients.
  - Corrected the data type for `ProcessInfo.Id` to `Double` in the Android client to align with the Hub's data type and prevent deserialization errors, with necessary casting for compatibility.

+++ OmniSync.Android/app/src/main/java/com/omni/sync/ui/screen/ProcessScreen.kt
@@ -249,7 +249,7 @@ fun ProcessItem(process: ProcessInfo, signalRClient: SignalRClient?) {
             // Compact Kill Button
             IconButton(
                 onClick = { 
-                    signalRClient?.killProcess(process.id)?.subscribe() 
+                    signalRClient?.killProcess(process.id.toInt())?.subscribe() 
                 },
                 modifier = Modifier.size(24.dp)
             ) {
